Import iCub model from icub-models repo
Only the URDF file has been imported. All the meshes and configuration files have
been not copied, and the model is still referring to the meshes and configuration files
installed by icub-models, that so is effectively a dependency for this repo.
If you want to modify a mesh or a configuration file, for import it in this repo,
and then modify the URDF file to refer to it specifying its path w.r.t. indicating that
it belongs to the icub_head_epr directory.

Commit reference: https://github.com/robotology/icub-models/tree/5bc520ea2bd8368422818badc6893c4c81201ea4/iCub_manual/robots/iCubGazeboV2_5_visuomanip